requirements/base.txt
django-environ
Django
pillow
python-telegram-bot
celery
django-jazzmin
django-rosetta
django-modeltranslation
whitenoise
djangorestframework
drf-yasg
django-cors-headers
django-allauth
django-auditlog
django-celery-beat==2.5.0
django-celery-results
django-ckeditor
django-debug-toolbar
django-filter
django-modeltranslation
django-phonenumber-field
django-recaptcha
django-redis
django-rosetta
django-timezone-field
djangorestframework-simplejwt
Faker
psycopg2
requests
redis
twilio
whitenoise
beautifulsoup4
phonenumbers==8.13.11
dj-rest-auth
